+++ lib/autokey/scripting/abstract_dialog.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2011 Chris Dekter
+# Copyright (C) 2023 sebastiansam55
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -12,40 +12,24 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
+Abstract interface for dialog interactions.
+This is an abstraction layer for platform dependent dialog handling.
+It unifies dialog handling for Qt, GTK and headless autokey UIs.
+"""
 
-"""Class for creating Qt dialogs"""
+from abc import ABC, ABCMeta, abstractmethod
+from pathlib import Path
 
-import subprocess
+logger = __import__("autokey.logger").logger.get_logger(__name__)
 
-from autokey.scripting.common import DialogData, ColourData
-
-from autokey.scripting.abstract_dialog import AbstractDialog
-
-class QtDialog(AbstractDialog):
+class AbstractDialog(ABC):
+    __metaclass__ = ABCMeta
     """
-    Provides a simple interface for the display of some basic dialogs to collect information from the user.
-
-    This version uses KDialog to integrate well with KDE. To pass additional arguments to KDialog that are
-    not specifically handled, use keyword arguments. For example, to pass the --geometry argument to KDialog
-    to specify the desired size of the dialog, pass C{geometry="700x400"} as one of the parameters. All
-    keyword arguments must be given as strings.
-
-    A note on exit codes: an exit code of 0 indicates that the user clicked OK.
+    Abstract interface for dialog interactions.
+    This is an abstraction layer for platform dependent dialog handling.
+    It unifies dialog handling for Qt, GTK and headless autokey UIs.
     """
-
-    def _run_kdialog(self, title, args, kwargs) -> DialogData:
-        for k, v in kwargs.items():
-            args.append("--" + k)
-            args.append(v)
-
-        with subprocess.Popen(
-                ["kdialog", "--title", title] + args,
-                stdout=subprocess.PIPE,
-                universal_newlines=True) as p:
-            output = p.communicate()[0][:-1]  # type: str # Drop trailing newline
-            return_code = p.returncode
-
-        return DialogData(return_code, output)
 
     def send_notification(self, title, message, icon="autokey", timeout="10", **kwargs):
         """
@@ -58,7 +42,7 @@ class QtDialog(AbstractDialog):
         @return: A tuple containing the exit code and user input
         @rtype: C{DialogData(int, str)}
         """
-        return self._run_kdialog(title, ["--title", title, "--passivepopup", message, "--icon", icon, str(timeout)], kwargs)
+        return
 
     def info_dialog(self, title="Information", message="", **kwargs):
         """
@@ -71,7 +55,7 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and user input
         @rtype: C{DialogData(int, str)}
         """
-        return self._run_kdialog(title, ["--msgbox", message], kwargs)
+        return
 
     def input_dialog(self, title="Enter a value", message="Enter a value", default="", **kwargs):
         """
@@ -85,7 +69,7 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and user input
         @rtype: C{DialogData(int, str)}
         """
-        return self._run_kdialog(title, ["--inputbox", message, default], kwargs)
+        return
 
     def password_dialog(self, title="Enter password", message="Enter password", **kwargs):
         """
@@ -98,7 +82,7 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and user input
         @rtype: C{DialogData(int, str)}
         """
-        return self._run_kdialog(title, ["--password", message], kwargs)
+        return
 
     def combo_menu(self, options, title="Choose an option", message="Choose an option", **kwargs):
         """
@@ -112,7 +96,7 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and user choice
         @rtype: C{DialogData(int, str)}
         """
-        return self._run_kdialog(title, ["--combobox", message] + options, kwargs)
+        return
 
     def list_menu(self, options, title="Choose a value", message="Choose a value", default=None, **kwargs):
         """
@@ -127,25 +111,7 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and user choice
         @rtype: C{DialogData(int, str)}
         """
-
-        choices = []
-        optionNum = 0
-        for option in options:
-            choices.append(str(optionNum))
-            choices.append(option)
-            if option == default:
-                choices.append("on")
-            else:
-                choices.append("off")
-            optionNum += 1
-
-        return_code, result = self._run_kdialog(title, ["--radiolist", message] + choices, kwargs)
-        if (return_code == 0):
-            choice = options[int(result)]
-        else:
-            choice = None
-
-        return DialogData(return_code, choice)
+        return
 
     def list_menu_multi(self, options, title="Choose one or more values", message="Choose one or more values",
                         defaults: list=None, **kwargs):
@@ -161,26 +127,7 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and user choice
         @rtype: C{DialogData(int, List[str])}
         """
-
-        if defaults is None:
-            defaults = []
-        choices = []
-        optionNum = 0
-        for option in options:
-            choices.append(str(optionNum))
-            choices.append(option)
-            if option in defaults:
-                choices.append("on")
-            else:
-                choices.append("off")
-            optionNum += 1
-
-        return_code, output = self._run_kdialog(title, ["--separate-output", "--checklist", message] + choices, kwargs)
-        results = output.split()
-
-        choices = [options[int(choice_index)] for choice_index in results]
-
-        return DialogData(return_code, choices)
+        return
 
     def open_file(self, title="Open File", initialDir="~", fileTypes="*|All Files", rememberAs=None, **kwargs):
         """
@@ -195,11 +142,8 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and file path
         @rtype: C{DialogData(int, str)}
         """
-        if rememberAs is not None:
-            return self._run_kdialog(title, ["--getopenfilename", initialDir, fileTypes, ":" + rememberAs], kwargs)
-        else:
-            return self._run_kdialog(title, ["--getopenfilename", initialDir, fileTypes], kwargs)
-
+        return
+    
     def save_file(self, title="Save As", initialDir="~", fileTypes="*|All Files", rememberAs=None, **kwargs):
         """
         Show a Save As dialog
@@ -213,10 +157,8 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and file path
         @rtype: C{DialogData(int, str)}
         """
-        if rememberAs is not None:
-            return self._run_kdialog(title, ["--getsavefilename", initialDir, fileTypes, ":" + rememberAs], kwargs)
-        else:
-            return self._run_kdialog(title, ["--getsavefilename", initialDir, fileTypes], kwargs)
+        return
+
 
     def choose_directory(self, title="Select Directory", initialDir="~", rememberAs=None, **kwargs):
         """
@@ -230,10 +172,7 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and chosen path
         @rtype: C{DialogData(int, str)}
         """
-        if rememberAs is not None:
-            return self._run_kdialog(title, ["--getexistingdirectory", initialDir, ":" + rememberAs], kwargs)
-        else:
-            return self._run_kdialog(title, ["--getexistingdirectory", initialDir], kwargs)
+        return
 
     def choose_colour(self, title="Select Colour", **kwargs):
         """
@@ -245,11 +184,7 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and colour
         @rtype: C{DialogData(int, str)}
         """
-        return_data = self._run_kdialog(title, ["--getcolor"], kwargs)
-        if return_data.successful:
-            return DialogData(return_data.return_code, ColourData.from_html(return_data.data))
-        else:
-            return DialogData(return_data.return_code, None)
+        return
 
     def calendar(self, title="Choose a date", format_str="%Y-%m-%d", date="today", **kwargs):
         """
@@ -265,4 +200,4 @@ class QtDialog(AbstractDialog):
         @return: a tuple containing the exit code and date
         @rtype: C{DialogData(int, str)}
         """
-        return self._run_kdialog(title, ["--calendar", title], kwargs)
+        return